+++ CHANGELOG.md
@@ -1,5 +1,28 @@
 # CHANGELOG
 
+## Release candidate
+
+### Breaking changes
+- Endpoint for updating custom tenant data changed (#98, PLUM Sprint 221104)
+
+### Fix
+- Logout with ID token (#116, PLUM Sprint 221104)
+- Disable registration service when no credential provider supports registration (#118, PLUM Sprint 221104)
+
+### Features
+- Roles have an optional "description" field (#103, PLUM Sprint 221021)
+- User registration (invitation only) (#86, PLUM Sprint 221021)
+- Delete and rename resources (#113, PLUM Sprint 221104)
+- List roles that contain a specific resource (#113, PLUM Sprint 221104)
+- Include session ID and parent session ID in ID token (#116, PLUM Sprint 221104)
+
+### Refactoring
+- Keep superuser role after provisioning (#102, PLUM Sprint 221021)
+- Endpoint for updating custom tenant data changed (#98, PLUM Sprint 221104)
+
+---
+
+
 ## v22.44
 
 ### Fix
@@ -14,6 +37,7 @@
 
 ### Refactoring
 - Simplified Admin UI client provisioning setup (#95, PLUM Sprint 221007)
+- Endpoint for updating custom tenant data changed (#98, PLUM Sprint 221007)
 
 ---